@@ -5,14 +5,22 @@ namespace App\Http\Controllers\Dashboard;
 use App\Http\Controllers\Controller;
 use App\Models\Role;
 use App\Models\User;
-use Illuminate\Validation\Rule;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('permission:users-create', ['only' => ['create','store']]);
+        $this->middleware('permission:users-read',   ['only' => ['show', 'index']]);
+        $this->middleware('permission:users-update',   ['only' => ['edit','update']]);
+        $this->middleware('permission:users-delete',   ['only' => ['delete']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -20,8 +28,8 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::paginate(5);
-        return view('dashboard.users.index',compact('users'));
+        $users = User::paginate(7);
+        return view('dashboard.users.index', compact('users'));
     }
 
     /**
@@ -31,8 +39,8 @@ class UserController extends Controller
      */
     public function create()
     {
-        $roles = Role::all();
-        return view('dashboard.users.create',['roles' => $roles]);
+        $roles = Role::get();
+        return view('dashboard.users.create', compact('roles'));
     }
 
     /**
@@ -44,24 +52,26 @@ class UserController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'name' => 'required|string|min:2',
+            'first_name' => 'required|string|min:2',
+            'last_name' => 'required|string|min:2',
             'email' => 'required|unique:users,email',
             'photo' => 'image',
             'password' => 'min:6',
-            'role_id' => 'required',
+            
         ]);
-        $data = $request->except('photo','password','role_id');
+        $data = $request->except('photo','password');
         if($request->hasFile('photo')){
             $path = $request->file('photo')->store(
-                'user','public_path'
+                'users','upload'
             );
             $data['photo'] = $path;
         };
         $data['password'] = Hash::make($request->password);
         $user = User::create($data);
-        $user->roles()->sync($request->role_id);
-        Alert::success(__('Congrates'), __('Added seccessfly'));
-        return redirect()->route('dashboard.users.index');
+        $user->syncRoles($request->roles);
+        $user->syncPermissions(DB::table('permission_role')->whereIn('role_id',$request->roles)->pluck('permission_id'));
+        Alert::success(__('Congratulations'), __('Added seccessfly'));
+        return redirect()->route('admin.users.index');
     }
 
     /**
@@ -81,11 +91,11 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(User $user)
+    public function edit($id)
     {
-        $roles = Role::all();
-
-        return view('dashboard.users.edit',['user' => $user,'roles' => $roles]);
+        $roles = Role::get();
+        $user = User::findOrFail($id);
+        return view('dashboard.users.edit',compact('user', 'roles'));
     }
 
     /**
@@ -98,27 +108,31 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         $request->validate([
-            'name' => 'required|string|min:2',
+            'first_name' => 'required|string|min:2',
+            'last_name' => 'required|string|min:2',
             'email' => [ 'required',Rule::unique('users')->ignore($user->id)],
             'photo' => 'image',
-            'role_id' => 'required',
+            'roles' => 'required'
         ]);
-
-        $data = $request->except('photo','role_id');
+    
+        $data = $request->except('photo');
+        if($request->password == null){
+            unset($data['password']);
+        }
         if($request->hasFile('photo')){
-            Storage::disk('public_path')->delete($user->photo);
+            Storage::disk('upload')->delete($user->photo);
             $path = $request->file('photo')->store(
-                'user','public_path'
+                'users','upload'
             );
             $data['photo'] = $path;
         } else {
             unset($data['photo']);
         }
         $user->update($data);
-        $user->roles()->sync($request->role_id);
-        Alert::success(__('Congrates'), __('Updated seccessfly'));
-        return redirect()->route('dashboard.users.index');
-       
+        $user->syncRoles($request->roles);
+        $user->syncPermissions(DB::table('permission_role')->whereIn('role_id',$request->roles)->pluck('permission_id'));
+        Alert::success(__('Congratulations'), __('Updated seccessfly'));
+        return redirect()->route('admin.users.index');
     }
 
     /**
@@ -127,14 +141,37 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request)
+    public function destroy(Request $request, $id)
     {
         $user = User::findOrFail($request->id);
-        if($user->photo != 'user/user.png'){
-            Storage::disk('public_path')->delete($user->photo);
+        if($user->photo != 'admin/admin.jpg'){
+            Storage::disk('upload')->delete($user->photo);
         }
         $user->delete();
-        Alert::success(__('Congrates'), __('Deleted seccessfly'));
-        return redirect()->route('dashboard.users.index');
+        Alert::success(__('Congratulations'), __('Deleted seccessfly'));
+        return redirect()->route('admin.users.index');
+    }
+
+    public function password()
+    {
+        return view('dashboard.profile.change_password');
+    }
+    
+    public function change_password(Request $request)
+    {
+        $request->validate([
+            'old_password' => 'required',
+            'new_password' =>  'required|confirmed'
+        ]);
+
+        if(!Hash::check($request->old_password, auth()->user()->password)){
+            return back()->with('error', __('Old password dose not match!'));
+        }
+
+        User::whereId(auth()->user()->id)->update([
+            'password' => Hash::make($request->new_password),
+        ]);
+        Alert::success(__('Congratulations'),__('Updated seccessfly'));
+        return redirect()->route('admin.change_password');
     }
 }
