@@ -2,22 +2,9 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
+use Laratrust\Models\LaratrustRole;
 
-class Role extends Model
+class Role extends LaratrustRole
 {
-    use HasFactory;
-    protected $fillable = [
-        'name',
-        'abilities',
-    ];
-    
-    protected $casts = [
-        'abilities' => 'array'
-    ];
-    public function users(){
-        return $this->belongsToMany(User::class,'role_user');
-    }
+    public $guarded = [];
 }
-
