@@ -2,14 +2,16 @@
 
 namespace App\Models;
 
-use Illuminate\Contracts\Auth\MustVerifyEmail;
+// use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
+use Laratrust\Traits\LaratrustUserTrait;
 
 class User extends Authenticatable
 {
+    use LaratrustUserTrait;
     use HasApiTokens, HasFactory, Notifiable;
 
     /**
@@ -18,10 +20,12 @@ class User extends Authenticatable
      * @var array<int, string>
      */
     protected $fillable = [
-        'name',
+        'first_name',
+        'last_name',
+        'photo',
+        'type',
         'email',
         'password',
-        'photo',
     ];
 
     /**
@@ -45,10 +49,8 @@ class User extends Authenticatable
 
     public function getImagePathAttribute()
     {
-        return asset('photo/'.$this->photo);
+        return asset('images/'.$this->photo);
             
     }
-    public function roles(){
-        return $this->belongsToMany(Role::class,'role_user');
-    }
+
 }
