@@ -14,8 +14,8 @@ class BroadcastServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        Broadcast::routes(['middleware' => 'auth:web,admin']);
-        
+        Broadcast::routes();
+
         require base_path('routes/channels.php');
     }
 }
