@@ -15,11 +15,13 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
+            $table->string('first_name');
+            $table->string('last_name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->string('photo')->default('user/user.png');
+            $table->string('photo')->default('admin/admin.jpg');
+            $table->string('type')->default('user');
             $table->rememberToken();
             $table->timestamps();
         });
