@@ -1,51 +1,57 @@
 @extends('layouts.dashboard.app')
+@section('title')
+@lang('Roles')
+@endsection
 @section('content')
-<table class="table table-striped table-hover">
-    <a class="btn btn-primary my-2" href="{{ route('dashboard.roles.create') }}">
-        <i class="bi bi-plus-lg"></i>
-        @lang('Add')
-    </a>
-    <thead>
+
+<table class="table theme role-table">
+  @permission('roles-create')
+    <a href="{{ route('admin.roles.create') }}" class="btn MyBtn mb-2"><span><i class="bi bi-plus"></i> @lang('Add')</span></a>
+  @endpermission
+    <thead class="">
         <tr>
-            <th>#</th>
-            <th>@lang('Name')</th>
-            <th>@lang('Users')</th>
-            <th>@lang('Control')</th>
+            <th>Name</th>
+            <th>Actions</th>
         </tr>
     </thead>
     <tbody>
-        @foreach($roles as $index=>$role)
+        @foreach($roles as $role)
         <tr>
-            <td>{{ $index+1 }}</td>
-            <td>{{ $role->name }}</td>
-            <td>{{ $role->users_count }}</td>
             <td>
-                <a href="{{ route('dashboard.roles.edit',$role->id) }} " title="Edit">
-                    <i class="bi bi-pencil-square bg-primary text-white p-1 rounded-1"></i>
-                </a>
-
-                <button href="" title="Delete" data-bs-toggle="modal" data-bs-target="#exampleModal" data-bs-userId="{{ $role->id }}">
-                    <i class="bi bi-trash3-fill bg-danger text-white p-1 rounded-1"></i>
-                </button>
-
+                <div class="d-flex align-items-center">
+                    <div class="ms-3">
+                        <p class="fw-bold mb-1">{{ $role->name }}</p>
+                    </div>
+                </div>
+            </td>
+            <td class="d-flex gap-2">
+            @permission('roles-show')
+                <a href="{{ route('admin.roles.show',$role->id) }}" class="btn btn-outline-success"><span><i class="bi bi-eye"></i></span></a>
+                @endpermission
+                @permission('roles-update')
+                <a href="{{ route('admin.roles.edit',$role->id) }}" class="btn btn-outline-primary"><span><i class="bi bi-pencil-square"></i></span></a>
+                @endpermission
+                @permission('roles-delete')
+                <button class="btn btn-outline-danger"
+                data-bs-toggle="modal" data-bs-target="#exampleModal" data-bs-roleId="{{ $role->id }}"
+                ><span><i class="bi bi-trash3"></i></span></button>
+                @endpermission
             </td>
         </tr>
         @endforeach
     </tbody>
 </table>
 
-<!-- Modal -->
-
 <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
   <div class="modal-dialog">
     <div class="modal-content">
       <div class="modal-body">
           <h2 class="text-center p-5">@lang('Are you sure to delete') ؟</h2>
-        <form action="{{ url('dashboard/users/destroy') }}" method="post">
+        <form action="{{ url('admin/roles/destroy') }}" method="post">
             @csrf 
             @method('delete')
           <div class="mb-3">
-            <input type="hidden" class="form-control" name="id" id="id">
+            <input type="hidden" class="form-control" value="" name="id" id="id">
           </div>
         
       </div>
@@ -57,8 +63,5 @@
     </div>
   </div>
 </div>
-{{ $roles->links() }}
+
 @endsection
-
-
-<!-- onclick="event.preventDefault();document.getElementById('delete').submit(); -->
