@@ -1,57 +1,55 @@
 @extends('layouts.dashboard.app')
 @section('content')
-<table class="table table-striped table-hover">
-    <a class="btn btn-primary my-2" href="{{ route('dashboard.users.create') }}">
-        <i class="bi bi-plus-lg"></i>
-        @lang('Add')
-    </a>
-    <thead>
+
+
+
+<table class="table align-middle theme">
+    @permission('users-create')
+    <a href="{{ route('admin.users.create') }}" class="btn MyBtn mb-2"><span><i class="bi bi-plus"></i> @lang('Add')</span></a>
+    @endpermission
+    <thead class="">
         <tr>
-            <th>#</th>
-            <th>@lang('Name')</th>
-            <th>@lang('email')</th>
-            <th>@lang('Photo')</th>
-            <th>@lang('Control')</th>
+            <th>Name</th>
+            <th>Actions</th>
         </tr>
     </thead>
     <tbody>
-        @foreach($users as $index=>$user)
+        @foreach($users as $user)
         <tr>
-            <td>{{ $index+1 }}</td>
-            <td>{{ $user->name }}</td>
-            <td>{{ $user->email }}</td>
             <td>
-                <img src="{{ $user->image_path }}" alt="{{ $user->name }}" class="img-thumbnail img-responsive">
+                <div class="d-flex align-items-center">
+                    <img src="{{ $user->image_path }}" class="rounded-circle" alt="" style="width: 45px; height: 45px" />
+                    <div class="ms-3">
+                        <p class="fw-bold mb-1">{{ $user->first_name.' '.$user->last_name }}</p>
+                        <p class="text-muted mb-0">{{ $user->email }}</p>
+                    </div>
+                </div>
             </td>
-            <td>
-                <!-- -->
-
-                <a href="{{ route('dashboard.users.edit',$user->id) }} " title="Edit">
-                    <i class="bi bi-pencil-square bg-primary text-white p-1 rounded-1"></i>
-                </a>
-
-                <button href="" title="Delete" data-bs-toggle="modal" data-bs-target="#exampleModal" data-bs-userId="{{ $user->id }}">
-                    <i class="bi bi-trash3-fill bg-danger text-white p-1 rounded-1"></i>
-                </button>
-
+            <td class="d-flex gap-2">
+                @permission('users-update')
+                <a href="{{ route('admin.users.edit',$user->id) }}" class="btn btn-outline-primary"><span><i class="bi bi-pencil-square"></i></span></a>
+                @endpermission
+                @permission('users-delete')
+                <button class="btn btn-outline-danger"
+                data-bs-toggle="modal" data-bs-target="#exampleModal" data-bs-userId="{{ $user->id }}"
+                ><span><i class="bi bi-trash3"></i></span></button>
+                @endpermission
             </td>
         </tr>
         @endforeach
     </tbody>
 </table>
 
-<!-- Modal -->
-
 <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
   <div class="modal-dialog">
     <div class="modal-content">
       <div class="modal-body">
           <h2 class="text-center p-5">@lang('Are you sure to delete') ؟</h2>
-        <form action="{{ url('dashboard/users/destroy') }}" method="post">
+        <form action="{{ url('admin/users/destroy') }}" method="post">
             @csrf 
             @method('delete')
           <div class="mb-3">
-            <input type="hidden" class="form-control" name="id" id="id">
+            <input type="hidden" class="form-control" value="" name="id" id="id">
           </div>
         
       </div>
@@ -63,8 +61,5 @@
     </div>
   </div>
 </div>
-{{ $users->links() }}
+{{ $users->withQueryString()->links() }}
 @endsection
-
-
-<!-- onclick="event.preventDefault();document.getElementById('delete').submit(); -->
