@@ -14,5 +14,5 @@ use Illuminate\Support\Facades\Broadcast;
 */
 
 Broadcast::channel('App.Models.User.{id}', function ($user, $id) {
-    return true;
-},['guards' => ['web', 'admin']]);
+    return (int) $user->id === (int) $id;
+});
